pyproject.toml
[tool.poetry]
name = "quip-api-es"
version = "0.1.0"
description = "API de frases en español (dataset curado + endpoints en FastAPI)."
authors = ["CoderDeltaLAN <coderdeltalan.cargo784@8alias.com>"]
license = "MIT"
readme = "README.md"
packages = [
  { include = "quip_api_es", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.115.0"
uvicorn = {version="^0.30.0", extras=["standard"]}
pydantic = "^2.7.0"
orjson = "^3.10.0"
pillow = "^10.3.0"
pytesseract = "^0.3.10"
python-slugify = "^8.0.0"
unidecode = "^1.3.8"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
black = "^24.4.0"
pytest = "^8.2.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

# ----- Herramientas -----
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E","F","I","UP"]
ignore = []
fixable = ["ALL"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]
